+++ ContactProcess/particle.h
@@ -7,9 +7,9 @@
 #include <random>
 class particle
 {
-    static std::vector <std::vector<int>> stensil;
+    static std::vector <std::vector<long int>> stensil;
     static std::set <std::vector<long int>> particle_locations;
-    static int number_of_particles;
+    static long int number_of_particles;
     static double lambda;
     static const double epsilon;
     static double spawning_range;
@@ -18,32 +18,32 @@ class particle
 public:
     std::vector<long int> location;
 private:
-    std::set <std::vector <int>> spawning_sites;
+    std::set <std::vector <long int>> spawning_sites;
     
 
 
 public:
 
-    void initialize_system_locations(std::set <std::vector<int>>);
+    void initialize_system_locations(std::set <std::vector<long int>>);
     
     int get_spawning_range() const;
     
-    int get_num_sites() const;
+    long int get_num_sites() const;
 
-    std::unique_ptr<particle>  spawn_particle(int &offset);
+    std::unique_ptr<particle>  spawn_particle(long int &offset);
     
     std::pair<std::unique_ptr<particle>,bool> poke_particle(int &rand, double prob);
-    static int get_num_particles();
+    static long int get_num_particles();
     static void reset_num_particles();
     
-    std::set <std::vector<int>> get_particle_locations() const;
+    std::set <std::vector<long int>> get_particle_locations() const;
 //    std::set <std::vector<int>> get_spawn_sites() const;
     
     void set_lambda(double);
     void set_R(size_t new_val);
     void set_dim(size_t new_val);
     
-    particle(std::vector<int> loc =  std::vector<int>(dimension,0));
+    particle(std::vector<long int> loc =  std::vector<long int>(dimension,0));
     ~particle();
     static void operator delete(void* ptr);
